requirements.txt
awscli==1.16.156
boto==2.49.0
boto3==1.9.134
botocore==1.12.146
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
gensim==3.4.0
matplotlib==3.1.0
nltk==3.4.1
numpy==1.16.2
pandas==0.24.2
PyMySQL==0.9.3
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
SQLAlchemy==1.3.3
statsmodels==0.9.0
urllib3==1.24.3
virtualenv==16.5.0
